+++ app/services/bridge/banks/list.rb
@@ -1,19 +1,17 @@
-class Bridge::Categories::ListAllCategories
+class Bridge::Banks::List
   require 'rest-client'
 
-  def self.call
-    url = "https://sync.bankin.com/v2/categories?limit=500"
+def self.call(access_token)
+    url = "https://sync.bankin.com/v2/banks"
     params = {
-      limit: 500,
       client_id: ENV["bridge_client_id"],
       client_secret: ENV["bridge_client_secret"]
     }
     response = RestClient.get(url, {
       params: params,
       "Bankin-Version": "2018-06-15",
-      "Accept-Language": "FR"
+      "Authorization": "Bearer #{access_token}"
     })
     JSON.parse(response.body)["resources"]
   end
 end
-
