+++ app/services/bridge/items/list.rb
@@ -1,8 +1,8 @@
-class Bridge::Users::ConnectItem
+class Bridge::Items::List
   require 'rest-client'
 
   def self.call(access_token)
-    url = "https://sync.bankin.com/v2/connect/items/add/url"
+    url = "https://sync.bankin.com/v2/items"
     params = {
       client_id: ENV["bridge_client_id"],
       client_secret: ENV["bridge_client_secret"]
@@ -12,6 +12,6 @@ class Bridge::Users::ConnectItem
       "Bankin-Version": "2018-06-15",
       "Authorization": "Bearer #{access_token}"
     })
-    response.code == 200 ? JSON.parse(response.body)["redirect_url"] : response
+    JSON.parse(response.body)["resources"]
   end
 end
