+++ app/services/bridge/items/refresh_status.rb
@@ -1,8 +1,8 @@
-class Bridge::Banks::ListBanks
+class Bridge::Items::RefreshStatus
   require 'rest-client'
 
-def self.call(access_token)
-    url = "https://sync.bankin.com/v2/banks"
+  def self.call(item_id, access_token)
+    url = "https://sync.bankin.com/v2/items/#{item_id}/refresh/status"
     params = {
       client_id: ENV["bridge_client_id"],
       client_secret: ENV["bridge_client_secret"]
@@ -12,6 +12,6 @@ def self.call(access_token)
       "Bankin-Version": "2018-06-15",
       "Authorization": "Bearer #{access_token}"
     })
-    JSON.parse(response.body)["resources"]
+    JSON.parse(response.body)
   end
 end
